Migration and Seeder slightly tampered; Blade as well.

+++ database/migrations/2017_09_08_025553_create_diseases_table.php
@@ -16,7 +16,7 @@ class CreateDiseasesTable extends Migration
         Schema::create('diseases', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->unique();
-            $table->unsignedInteger('weight');
+            $table->double('weight',1,1);
             $table->unsignedInteger('disease_type_id');
             $table->unsignedInteger('created_by')->nullable();
             $table->unsignedInteger('updated_by')->nullable();
